added setAccessibilityDelegate in adapter
